network: move FilterPodNetworkRoutes away from the driver pkg
This helper function was moved to the `link` pkg.
