+++ pkg/network/driver/common.go
@@ -443,22 +443,3 @@ func (h *NetworkUtilsHandler) DisableTXOffloadChecksum(ifaceName string) error {
 // Allow mocking for tests
 var DHCPServer = dhcpserver.SingleClientDHCPServer
 var DHCPv6Server = dhcpserverv6.SingleClientDHCPv6Server
-
-// filter out irrelevant routes
-func FilterPodNetworkRoutes(routes []netlink.Route, nic *cache.DHCPConfig) (filteredRoutes []netlink.Route) {
-	for _, route := range routes {
-		log.Log.V(5).Infof("route: %s", route.String())
-		// don't create empty static routes
-		if route.Dst == nil && route.Src.Equal(nil) && route.Gw.Equal(nil) {
-			continue
-		}
-
-		// don't create static route for src == nic
-		if route.Src != nil && route.Src.Equal(nic.IP.IP) {
-			continue
-		}
-
-		filteredRoutes = append(filteredRoutes, route)
-	}
-	return
-}
